finish imolement app on Realm data base

+++ planIt/Views/ListView.swift
@@ -9,37 +9,48 @@ import SwiftUI
 import RealmSwift
 
 struct ListView: View {
-    @StateObject var listViewModel : ListViewModel 
+    @EnvironmentObject var listViewModel : ListViewModel
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items ) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
-                    }
+        ZStack{
+            if listViewModel.items.isEmpty{
+                NoItemView()
+            
+            }else{
+                List{
+                    ForEach(listViewModel.items ) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
 
-            }
-            .onDelete(perform: { indexSet in
-                for index in indexSet{
-                   let item = listViewModel.items[index]
-                    listViewModel.deleteItem(item: item)
+                    }
+                    .onDelete(perform: { indexSet in
+                        for index in indexSet{
+                           let item = listViewModel.items[index]
+                            listViewModel.deleteItem(item: item)
+                        }
+                    })
+        //            .onMove { indexSet, newIndex in
+        //                listViewModel.moveItem(from: indexSet, to: newIndex)
+        //            }
+                    
+                    
                 }
-            })
-            .onMove { indexSet, newIndex in
-                listViewModel.moveItem(from: indexSet, to: newIndex)
+           
             }
-            
-            
         }
-        
         .listStyle(.plain)
         .navigationTitle("plan It 📝")
         .navigationBarItems(leading: EditButton(), trailing:
         NavigationLink("add", destination: AddView())
         )
+       
+          
+        
+       
 
     }
    
@@ -48,11 +59,11 @@ struct ListView: View {
 
 struct ListView_Previews: PreviewProvider {
     static var previews: some View {
-        let realm = try! Realm()
+//        let realm = try! Realm()
         NavigationView{
-            ListView(listViewModel: ListViewModel(realm: realm))
+            ListView()
         }
-        .environmentObject(ListViewModel(realm: realm))
+        .environmentObject(ListViewModel())
     }
 }
